requirements.txt
blinker==1.4
click==6.7
Flask==0.12.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
SQLAlchemy==1.1.15
Werkzeug==0.12.2
